services/tms/requirements.txt
Django>=5.1
djangorestframework>=3.15
django-cors-headers>=4.4
psycopg2-binary>=2.9
django-environ>=0.11
drf-spectacular>=0.27
celery>=5.4
redis>=5.0
django-redis>=5.4
djangorestframework-simplejwt>=5.3
django-filter>=24.3
requests>=2.32
python-jose[cryptography]>=3.3
